Menu Application
Utilizing a controlling variable, a do while loop is used with a switch statement to allow a menu that responds to user input. It repeats and prompts the user to select another option until case 5 where the controlling variable condition is changed to break out of the loop and end the application.